CUDA: fix double-free when double-waiting
Waiting multiple times on the same even results in the command's
finalization being run twice, which in turn results in a double call to
pocl_ndrange_node_cleanup for the same command, and finally a
double-free.

The solution is simply to not finalize the command again if it has
finished already.