pocl_timing: fix detection of Unix monotonic clocks
CLOCK_UPTIME_FAST is an extension only available on FreeBSD and
DragonFlyBSD, so don't try to use it on other BSDs - this previously
resulted a compilation failure on NetBSD.

Instead, use CLOCK_MONOTONIC. This is available on NetBSD and
OpenBSD and is defined in POSIX.2008. On older Unix-like
operating systems, keep the fallback to CLOCK_REALTIME.

Replace OS-specific checks with more portable checks for definitions.

Signed-off-by: Nia Alarie <nia@NetBSD.org>