Add a fallback to finding the device's *.so
On linux, a user application can find a *.so with the
LD_LIBRARY_PATH evironment variable.

To use the LD_LIBRARY_PATH, the path returned by
get_pocl_device_lib_path() needs to be just the file name.
So add option to do this.

In pocl_init_device() if the current method fails, try again
with the fallback.

Signed-off-by: Tom Rix <trix@redhat.com>