Partly revert & fix commit 4d400e7a54f0e
The "Fix compilation errors when building on macOS" commit
was already partly reverted in commit 010845ab0f6d95e7ec267.

The fix is incorrect, OPENCL_LIBS is already set to either
libOpenCL or libpocl, depending on value of ENABLE_ICD.

The problem is that bin/poclcc is built against libpoclu,
which references clCreateProgramWithIL (which is 2.0 API),
and this ends up as symbol in poclcc.

Instead this removes clCreateProgramWithIL from bin/poclcc.