Specialized work-group functions in PoCL kernel program binaries
By default the PoCL program binaries store generic kernel binaries
which can be executed across any grid dimensions. This configuration
variable can be used to also include specialized work-group functions
in the binaries, by  defining a comma separated list of strings that
describe the specialized versions. The strings adhere to the directory
names in the PoCL cache from which the binaries are captured.