cuda : generalize pocl_cuda_submit_memfill
The use of cuMemsetD32|16 is dependent on the alignment and
as well as well as the size.  Check the alignment of the inputs
and use an appropriate cuMemset.

Handle the general size by breaking the memset into
the equivelent pocl_cuda_* calls.  first copy the pattern
from the host to the device memory, then copy pattern from
device to device memory.

Move pocl_cuda_sumit_memfill to after the calls it references.

Signed-off-by: Tom Rix <trix@redhat.com>