SVM API call fixes
According to spec, they should return CL_INVALID_OPERATION
if none of the devices are SVM capable

Implement a simple linked list of svm pointers and sizes in
the context. This is required for clCreateBuffer() with host_ptr
being a SVM pointer, and clGetMemObjectInfo().

CTS 3.0 SVM tests should now pass