Set platform to 3.0 and CPU device to 3.0 (.. on some conditions)
Only sets CPU device to 3.0 if LLVM version is >= 14.

Also changes the default of ENABLE_CONFORMANCE CMake variable to OFF.

The purpose remains the same, but now if ENABLE_CONFORMANCE=ON it
disables SPIR and SPIR-V because the new CTS tests SPIR and SPIR-V,
and PoCL's SPIR(-V) support is not yet good enough to pass.