add a workaround for conformance_api_ocl_30 / consistency_il_programs
The test is supposed to test that devices not supporting IL will
return CL_INVALID_OPERATION on clCreateProgramWithIL &
clSetProgramSpecializationConstant. But the test does this by
using an invalid IL / program. The Spec says that it is OK to
return CL_INVALID_VALUE if the IL is invalid, but the test
rejects that return value. Probably assumes that IL validity
is checked after device availability, but that's not true
for PoCL.