Builtin_kernels: move definitions to own files
This commit is cherry-picked from:
b86e34b2ddf7f4ccbd961fb13107a0ba67ab0915 .

The idea is to have a set of common kernels that work on all devices.
In order to do this, lib/CL/devices/builtin_kernels.{cc,hh} have been
created to store definitions and implementations.
This is related to the accel and almaif drivers, see these to get a
better idea.