opaque pointer getElementType deprecation
The use of getElementType method as started to be deprecated
because of opaque pointers in llvm 14 with their planned
removal in llvm 16.  So it is necessary to transition them now.

See the migration instructions here
https://llvm.org/docs/OpaquePointers.html

For AllocInst, since the replacement getAllocatedType is
available back to llvm 6, do a simple text replacement of
getType()->getElementType() to getAllocatedType()

For Globals use getValueType and define it as a macro
for legacy llvm's to reduce the if-def clutter.

Arguements was difficult and needs if-def's to call through
the function iterator to use getParamByValType.

Signed-off-by: Tom Rix <trix@redhat.com>