handle llvm 14 LLVMBuildGEP deprecation
Starting in llvm 14 LLVMBuildGEP is deprecated in favor of
LLVMBuildGEP2.  LLVMBuildGEP2 is nearly the same as LLVMBuildGEP.
The 2nd parameter in LLVMBuildGEP2 is a explicit type which used
to be implicitly found by the pointer parameter.

For pre llvm 14 use a macro to drop the type parameter and call
the LLVMBuildGEP.

Signed-off-by: Tom Rix <trix@redhat.com>