Fix a bug in SPIR-V caching
pocl_regen_spirv_binary() was writing the unfinished (not linked) binary
that came out from llvm-spirv into program.bc; this was later rewritten
by pocl_llvm_link_program() with the final linked binary. This was
unnecessary and also incorrect because if two threads in a program
try to build the same SPIRV, one could end up accessing the intermediate
program.bc and assume it is the final. This is fixed to only write
temporary files in pocl_regen_spirv_binary() and only write
the final linked program.bc in pocl_llvm_link_program().