TCE: fix consequent launches with different global offset
Take global offset into account when determining whether
the kernel is the same as the one already loaded on to the
device.

TODO: still a bit fragile to the kernel address being the same.
(Kernel address comparison can break if the kernel is freed
and a new kernel is created with same local size and global offset,
and it happens to have the same pointer value.)