lib/llvmopencl: add a program.bc pass to deal with program-scope variables
This is run on program.bc; it creates an initializer kernel for
program-scope variables, and replaces references of program-scope
variables in the kernels with references to extern buffer holding
the variables + an offset into the buffer. This extern buffer variable
is then replaced in the Workgroup pass by a load from pocl_context.