ProgramScopeVariables.cc: fix missing AddressSpace casts
If the program input is OpenCL C source, Clang uses Target AS
numbers in the IR. However if the input is SPIR-V, the SPIR AS
numbers are used. This affects program scope variables too.
The single buffer which acts as storage for all global
variables is created with Target global AS. Therefore
program scope variables in global AS coming from SPIR-V
need an AS cast, when replaced by loads from single
storage buffer.