Retain the event to be notified about in pocl_broadcast
This is required, since the driver's notify-function might
otherwise destroy the event.

This was manifesting as a race condition in almaif-driver,
which can execute the commands directly in its notify-function,
recursively processing the event wait lists. In some situations
it would destroy the event that was being notified about prematurely.