Refactor half/double CMake support & tests
Refactors old code that used CL_DISABLE_HALF and similar variables
to check for half/double support. This made sense when only CPU driver
existed, but we now have multiple drivers and only some support
half/double, so this needs to be handled per-device, and needs
to be done at runtime.

- removed ENABLE_FP64, CL_DISABLE_HALF, CL_DISABLE_DOUBLE
- HOST_CPU_SUPPORTS_FLOAT16 indicates _Float16 type support
- HOST_CPU_ENABLE_CL_KHR_FP16 indicates cl_khr_fp16 support on CPU
- HOST_CPU_ENABLE_CL_KHR_FP64 indicates cl_khr_fp64 support on CPU
- remove CL_DISABLE_X conditionals from tests, instead check
  for cl_khr_fpX at runtime and return skip code if it's not available