CPU: Replace OpenCL as cast builtins with addrspace casts
Special handling for the AS cast built-ins: They do not
type mangling, which complicates the CPU implementation
which sometimes gets all-0 AS input and sometimes not (SPIR-V).
Here, in case the target doesn't define these built-ins
in the bitcode library, we replace the calls directly with
address space casts. This works for uniform address space
targets (CPUs), while the disjoint AS targets can override the
implementation as needed. Fixes #1203.