Make locking more explicit in clRelease*
POCL_RELEASE_OBJECT locks the object, decrements the reference count and
then unlocks the object. This is fine as far as the reference count
itself is considered. All these functions had the rough form of:

POCL_RELEASE_OBJ (obj, new_refcount);
POCL_MSG_PRINT_REFCOUNTS ("%d", obj->id);
if (new_refcount == 0) { free(obj); } else { ... }

This causes a race condition:
1. T1 decrements the refcount
2. T2 gets scheduled instead of T1
3. T2 decrements refcount
4. T2 prints the refcount message and notices that new_refcount is zero
5. T2 releases all memory associated with obj and returns
6. T1 gets scheduled back
7. T1 attempts to print the refcount message but obj->id is now invalid

In order to fix the problem and signal what's going on to avoid repeating it,
the locking and unlocking is now done explicitly and the critical region is
extended into the if..else block.