clGetProgramInfo: return correct number of binaries
As per OpenCL specification V3.0.14, Table 29 in section 5.8.9.
"Program Object Queries" states that CL_PROGRAM_BINARY_SIZES and
CL_PROGRAM_BINARIES queries return an array containing an entry for
each *associated* device even if the program was not built for that
device. Furthermore the specification of clCreateProgramWithSource
(Section 5.8.1 "Creating Program Objects") states that *all* devices
associated with the given context are associated with the newly created
program.

Since progam->num_devices / program->devices only lists the devices for
which the program has actually been built, this list can't be used in
these queries.