pocl-d: Memleak fixes
Due to the reconnect support not having a timeout nor there are any
freeing of resources together with client-side context free, only free
(some) resources when reconnect is not enabled and when all connections
that refer to the same vcontext on server side are gone.

Most importantly, free the SVMPool if CG SVM is enabled. It can take a
good chunk of VM (e.g. 4 GiB) and leads quickly to OOM if not freed.

Spotted an issue with peer handling, which crashes when trying to
reset/free it. Something with the thread (to discuss with the orig.
author). For now, just leak it (it was leaking anyhow so far).

Future work:

- Time window for the ability to reconnect (a time limit to retain
context even though all connections using it are dropped)
- Try to actually share the shared context across multiple client side
contexts to avoid reinitializing some resources such as the SVM pool