pocld: Add some bounds checking in PoclDaemon
Merging the various vectors holding per-socket data does not work out
well since `poll()` expects a continuous array of only `struct pollfd`.
So let's just do bounds checking instead. The overhead should be
negligible (not in core parts of tight loops) and an optimizing
compiler should notice that some of the checks are redundant and
eliminate them (I added them in the code anyway for consistency).