PoCL-R: Attempt to allocate more SVM memory
Instead of a single maximum allocation, try to allocate as many times
as clSVMAlloc() returns a nullptr. The allocated regions might have
gaps for which there is a maximum threshold, depending on the luck
we have with the driver's memory allocator. At least for my Intel iGPU
this allows to SVM-allocate most of the global memory.

In the host-side the region must be allocated as a continuous area from
the minimum to the maximum allocated address (including the
unallocatable gaps).