Upgraded dpcpp-book-samples to its latest HEAD
Made the dpcpp book samples test CPU-specific for now. There was the
same problem as with the oneAPI-samples that the LD_LIBRARY_PATH
was overwritten, hiding ocl-icd, easily falling to the oneAPI's
OpenCL and thus not testing PoCL. Now the ocl-icd path is prepended
explicitly to ensure we direct to libpocl.so via ICD.

Made the build failable with -ik so we can skip the one that
requires PSTL from later oneAPIs.