Sub-buffer rework
Use the sub-buffer like it was a regular buffer wherever feasible
instead of pretending it is the parent buffer. This enables
fine-grain implicit migrations of the sub regions.

Add hooks to driver layer to be able to allocate resources
for the sub-buffers.

Also cleanup/refactor/comment the migration code a bit more.