Multidevice command buffering fix
The program_device_i was not copied, causing all kernel_arg objects
point to the 0th instance.

This caused a crash when there are two (or more) remote devices which
execute the same kernel: The remote driver, after executing K1 on D0,
thought that it has set the kernel arguments already when executing K1
on D1 although it had set it only to D0 (due to the wrongly shared
program_device_i==0).