Level Zero: disable shared memory use on dGPU
This is necessary to make CTS math tests (and subbuffers in general)
work on dGPU. The problem: it seems zeContextMakeMemoryResident() / Evict
ignore the size argument, and always make the whole buffer (un)resident.
This makes it impossible to have multiple CPU threads reliably working
with subbuffers of a single parent buffer.

In case we detect dGPU, disable sharedMem and use deviceMem
(with memcpy for map/unmap)