SPIR-V parser: implement a workaround for a bug in llvm-spirv translator
For OpAtomicCompareExchange(Weak), the translator sometimes inserts
OpBitcasts of the pointer argument to ptr_workgroup_uchar or similar
type, and this confuses the Level Zero driver. It seems the driver
uses the function argument types to infer the actual OpenCL
builtin name, and ends with non-existant builtin because of the uchar.

Workaround: change the bitcast to use pointer-to-return-type of the OpAtomic.
CTS test "c11_atomics" now passes with "atomic_compare_exchange_weak"
using Level Zero driver.