dbk: add JPEG de-/en-code kernels
This commit adds the JPEG DBK encode and decode
kernels and the infrastructure needed to run
kernels with a hidden state.
The JPEG encoding is done with libjpeg-turbo.
It requires the tj api level 3.0+ so this might
need to be built from source.
If cmake can not find your source built tj lib,
set the -Dlibjpeg-turbo_DIR cmake option to the
appropriate dir containing cmake files.

There is a test program that does an psnr
comparision after the image is encoded and
decoded by the DBKs.
Test data image is CC0 and provided by
Julius Ikkala