Run lcssa explicitly drop instcombine
LCSSA is ran to ensure it has generated its lcssa phis before
we break them in phistoallocas for the WILoop analysis. This is
an intermediate solution while working towards processing
unoptimized Clang output directly.

With this we do not need to run instcombine anymore, which
caused problematic CFG corner cases.