Ensure we always have a unique entry block in the PR
Ensure we always have a unique entry block in the PR where we
can add context restores. We should minimize code in that
region to avoid unnatural loops, but having the block there
is convenient during transformations since we always have
a known safe location where to inject context restores.