Handle cond barriers with CBS
Do not handle kernels with non-uniform predicates with WILoops
(for now). These used to be handled by "peeling" the first iteration of
the work-item loop to check if the barrier is taken or not, but it
complicates the parallel region formation and control flow construction
quite a bit without known benchmark benefits. The produced peeled loops
are also not easily vectorizable, thus the performance won't be good
anyhow for SIMD.

For VLIW/ILP the peeling might be useful though, but it still doesn't
justify the complexity add.

Remove peeling as well as explicit unrolling code to simplify
the WILoops handling a lot.