Proxy: add SPIR-V support
This commit adds the required infrastructure
needed to use clCreateProgramWithIl with a Proxy
device. It makes use of the SPIR-V parser for
metadata, so it can be used on platforms that do
not provide kernel arg metadata but do support
SPIR-V like the Android Mali GPU OpenCL libraries.
There has been a conversion of proxy.c->proxy.cpp
in order to use the parser and the mapping of
the parsed spir-v to pocl_kernel_metadata_t
from the level0 device has been moved and is now
shared with the proxy.