poclu: change spir-v program loading
This commit allows spir-v programs to be created
without having to have the cl_khr_il_program
extension available as long as it is above
version 2.1. This is more inline with the spec.
If this fails, clCreateProgramWithILKHR is called
instead.
This commit also adds some argument checking and
version displaying to example0.