set a stack-size-based WG size limit (CL_KERNEL_WORK_GROUP_SIZE) for CPU drivers
Adds an estimation of stack size (this is done at linking time,
on unoptimized bitcode, so the estimation is worst-case scenario).
The estimate is then used to set per-device & per-kernel value
of CL_KERNEL_WORK_GROUP_SIZE of each kernel in the program.

Stack overflow was causing failure of "conformance_main_generic_AS"
CTS test when run in debug build, or with Addr/Thread sanitizers enabled.