argbuffer wrapper: Always align args to MAX_EXTENDED_ALIGNMENT
The original HSA spec aligned the args to their size, but
it is problematic with structs of which sizes and alignment
can vary widely. Simplify the handling for now with the expense
of HSA-compliance (the HSA target is deprecated and unused nowadays)
and more argument buffer wasted.

Now the structs are also aligned to the max extended alignment,
thus should be able to contain max size OpenCL vectors as attributes,
adhering to the common struct alignment of aligning to the largest
aligned member.