msvc-ci: download the LLVM dep instead of building it
It as fast as using the cached LLVM without the cache miss penalty
plus saves job cache space.  A tradeoff to this is that PoCL won't be
tested on LLVM-18 as the LLVM binary package with libraries and
llvm-config for Windows is not available for it or the earlier
versions.