MSVC: Changes to support BUILD_SHARED_LIBS=ON
To enable this, `__declspec(dllexport)` needs to be placed on the
OpenCL API declarations too as MSVC requires matching attributes on
declarations and their definitions. This is done by injecting
`__declspec(dllexport)` into OpenCL headers by using CL_API_ENTRY
macro which is defined in pocl_cl.h for MSVC and BUILD_SHARED_LIBS=ON.

Because of this arrangement, direct `CL/*` includes under the `lib/` are
replaced with `pocl_cl.h` inclusion.