events: unlock notifier event on failed event
When a device notifies that an event failed,
the wait event eventually ends up in
pocl_update_event_finished. This function will
lock the command queue. But since the notify
event is locked in common.c:broadcast, a lock
order violation can occur as the command queue is
locked after the notifier event. This is in
contrast to the lock order in other places where
where it is first cq then event.
This commit solves this by unlocking the notifier
event before calling pocl_update_event_failed.