Fix issue #1649
Add a hack to always build pocld with only pocl_thread_c.{c,h}.

The proper way would be to fix the mess in pocld/CMakeLists
where it includes a bunch of out-of-project sources (../lib/CL/*)
and adds a bunch of other hacks to build them as part of pocld
sources. This should be separated into its own library & CMake
target which pocld should then link against.