Fix UnreachableToReturns when having uncond branches leading
The problem as the recursive call which invalidates the pred
iterator.  I moved the recursion logic to rely on the set and
repetitive calls from the layer above.  Also the deletion logic
now deletes from the leafs first to avoid having uses of
values produced by the earlier basic blocks.

Also cleaned up the code a bit while touching it.

Fixes the WILoops crash occuring with #1871 which was caused
by unreachables leaking to the WILoops processing which doesn't
expect to see them.