A non-robust work-around for #1887
This fixes an issue seen in #1887 but there is still possibility of the bug re-occurring. The reason is that that the `$TEMP` environment variable points into user's home directory too (by default) but in different form which in my case is `C:\Users\HENRYL~1\AppData\Local\Temp`. This works for me but might fail for other users in case there are "wide" characters appearing in very beginning of the home directory name.