Fixed a rematerialization case
An alloca was loaded in region A and the
loaded value was used in region B. In this case we should
remat at region B, but the code remat at the load site A,
thus the user in B could get corrupted data as it referred
to the value created in the other WI loop.

Improve remat to remat at the load value usage site
instead.

It appeared in a test case of #1886, which is attached to
this commit as a new test case.

Also improved the debug output for remat a bit.