Windows: fix search for HWLOC library (without pkg-config)
fixes the find_library(HWLOC) to also search for DLL,
create an IMPORTED target, and set its IMPORT_LOCATION property.

This allows us to use install(IMPORTED_RUNTIME_ARTIFACTS) to install
the hwloc DLL and ship it to the user.

Using pkg-config here would not help, because it does not set up
the IMPORT_LOCATION property.