The workaround for reduced fast flags not needed for LLVM 18+
721733862ad2 introduced a commit to not pass -cl-fast-relaxed-math
directly to Clang as it caused to inaccurate divisions in some cases.
However, the passed flags do not introduce essential relaxation
'reassoc' which is required for generating vectorized reduction
trees from float reduction loops.

The rounding problems seem to be fixed starting LLVM 18, thus we can
disable the workaround for newer LLVMs.