add_test_pocl to add 'cpu' label by default
...this is to avoid forgetting to add the label when
adding new tests.

'cpu' label means that the test is expected to pass
with host CPUs.

Add a new label 'xfail' that should be added to tests
that are known broken at the moment.

I don't think we need test-type labels (kernel, workgroup)
since the tests have the type in their name and can be
filtered by it.