WIL: fall back to CBS with unreachables again
For WIL currently we can only delete BBs with 'unreachable'
instructions in them to produce clean control flow graphs for the
loop-based methods. However, deletion of the whole basic
block (and redirecting the control to another one) is not correct when
relying on 'unreachable' to equal to 'abort', like Julia does.
Therefore, disable this behavior by default, making PoCL to fall back to
CBS when encountering unreachable instructions. Keep this environment
variable undocumented until there's time to look for a more robust and
optimizable solution. See, for example the Github Issue #1971 and a
bunch of others.

Also allow CBS to be ran on kernels without barriers. It should
work at least after canonbarriers which should add implicit entry/exit
barriers.