Command buffers: Allow mutable updates to pending cmd bufs
This makes the clUpdateMutableCommandsKHR readiness check match the
check in clEnqueueCommandBufferKHR, so that mutable updates are allowed
even if the cmd buf is still pending (and the cmd buf is simultaneous).

The check is refactored into an utility function, as it is now the same
in these two files.

Relevant Spec portions:
For clEnqueueCommandBufferKHR:
"CL_INVALID_OPERATION if command_buffer was not created with the
CL_COMMAND_BUFFER_SIMULTANEOUS_USE_KHR flag and is in the Pending
state."

And for the clUpdateMutableCommandsKHR:
"Using this function when command_buffer is in the pending state and not
created with the CL_COMMAND_BUFFER_SIMULTANEOUS_USE_KHR flag causes
undefined behavior."

So I think both imply that the operation should succeed when cmd buf is
in Pending state and was created with the SIMULTANEOUS-flag.