Allow bypassing the LLC_HOST_CPU autodetection
For some reason OpenASIP's LLVM doesn't list the host although
it has X86 support built in when installed with its installer
script. This allows to build with a CPU host using explicit
-DLLC_HOST_CPU=skylake (or whatever uarch you have).

My guess is that llc --version doesn't list the host in case
there's only one target built in.